Move non public API headers from FWCore/Framework/interface
The headers moved were not used by any other header and were only
used in the src/ or test/ sub-directories of FWCore/Framework
