+++ FWCore/Framework/src/OutputModuleDescription.h
