+++ FWCore/Framework/src/ProductDeletedException.cc
@@ -13,7 +13,7 @@
 // system include files
 
 // user include files
-#include "FWCore/Framework/interface/ProductDeletedException.h"
+#include "FWCore/Framework/src/ProductDeletedException.h"
 
 using namespace edm;
